backend/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
psycopg2-binary
greenlet==3.1.1
langgraph
langchain
groq
